Start structuring type helpers
