+++ sdk/src/_composables/authentication.ts
