+++ sdk/src/types/index.ts
@@ -1,3 +1,2 @@
-export type * from './command.js';
-export type * from './filter.js';
 export type * from './query.js';
+export type * from './schema.js';
