+++ sdk/src/types/schema.ts
@@ -1,0 +1,1 @@
+export type Relational<TItem extends object> = string | TItem;
