+++ sdk/src/types/utils.ts
@@ -1,0 +1,3 @@
+export type KeysOfType<O, T> = {
+	[K in keyof O]: O[K] extends T ? K : never;
+}[keyof O];
