--- sdk/src/utils/index.ts
@@ -1,1 +1,0 @@
-export { serializeParams } from './serialize-params.js';
